# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aphrodite API",
    "description": "API Description",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development Environment"
    },
    {
      "url": "http://test.aphrodite.com",
      "description": "Test Environment"
    }
  ],
  "paths": {
    "/v1/login": {
      "post": {
        "tags": [
          "Authentication Module"
        ],
        "description": "User login interface",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully retrieved user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResp"
                }
              }
            }
          }
        }
      }
    },
    "/v1/logout": {
      "post": {
        "tags": [
          "Authentication Module"
        ],
        "description": "User logout interface",
        "operationId": "logout",
        "responses": {},
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/send-code": {
      "post": {
        "tags": [
          "Authentication Module"
        ],
        "description": "Send verification SMS",
        "operationId": "send_verify_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendVerifyCodeReq"
              }
            }
          },
          "required": true
        },
        "responses": {}
      }
    },
    "/v1/user": {
      "get": {
        "tags": [
          "User Module"
        ],
        "description": "User information interface",
        "operationId": "info",
        "responses": {},
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "User Module"
        ],
        "description": "User update interface",
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserReq"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "User Module"
        ],
        "description": "User delete interface",
        "operationId": "delete",
        "responses": {},
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "GetUserResp": {
        "type": "object",
        "required": [
          "user_code",
          "user_no",
          "nickname",
          "email",
          "phone"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "nickname": {
            "type": "string",
            "example": "john"
          },
          "phone": {
            "type": "string",
            "example": "13800138000"
          },
          "user_code": {
            "type": "string",
            "example": "S8000"
          },
          "user_no": {
            "type": "integer",
            "format": "int64",
            "example": "100000",
            "minimum": 0
          }
        }
      },
      "LoginReq": {
        "type": "object",
        "required": [
          "phone",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "1234"
          },
          "phone": {
            "type": "string",
            "example": "13800138000"
          }
        }
      },
      "LoginResp": {
        "type": "object",
        "required": [
          "access_token"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "SendVerifyCodeReq": {
        "type": "object",
        "required": [
          "phone"
        ],
        "properties": {
          "phone": {
            "type": "string",
            "example": "13800138000"
          }
        }
      },
      "UpdateUserReq": {
        "type": "object",
        "required": [
          "nickname",
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "nickname": {
            "type": "string",
            "example": "john"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}